pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-bench"
version = "0.1.0"
description = "Benchmark LLM agents across providers on speed, cost, and quality."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "arcane-bear" }]
requires-python = ">=3.10"
keywords = ["llm", "benchmark", "agent", "evaluation", "openai", "anthropic", "gemini", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
openai = ["openai>=1.40"]
anthropic = ["anthropic>=0.39"]
gemini = ["google-generativeai>=0.8"]
ollama = ["ollama>=0.3"]
all = [
    "openai>=1.40",
    "anthropic>=0.39",
    "google-generativeai>=0.8",
    "ollama>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
agent-bench = "agent_bench.cli:main"

[project.urls]
Homepage = "https://github.com/arcane-bear/agent-bench"
Issues = "https://github.com/arcane-bear/agent-bench/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_bench"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]
